requirements.txt
pydantic
pytest
python-dotenv
openai
instructor